misc: refactored and renamed files
