+++ bin/_cli.js
@@ -1,3 +1,4 @@
+#!/usr/bin/env node
 /*
 	MIT License http://www.opensource.org/licenses/mit-license.php
 	Author Tobias Koppers @sokra
