+++ bin/getConfigFiles.js
